requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.1
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
executing==2.0.1
fastapi==0.110.2
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpie==3.2.2
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
orjson==3.10.1
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
Pygments==2.17.2
PySocks==1.7.1
pytest==8.1.1
pytest-cov==5.0.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
traitlets==5.14.3
types-pyasn1==0.6.0.20240402
types-python-jose==3.3.4.20240106
types-requests==2.31.0.20240406
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
